ilb/test: Set corrent listen port in backendPool
In the case of --mode=single-node, we do +1 for backend listen ports.
However, the +1 was not considered when creating backendPools. This led
to the same backend being used in most of the test cases.

The hidden +1 logic is going to be made explicit in a follow-up PR.

Reported-by: Marco Hofstetter <marco.hofstetter@isovalent.com>

+++ enterprise/test/ilb/helpers_k8s_resources.go
@@ -386,11 +386,11 @@ func lbService(namespace string, name string, opts ...serviceOption) *isovalentv
 
 type backendPoolOption func(o *isovalentv1alpha1.LBBackendPool)
 
-func withBackend(ip string, port int32) backendPoolOption {
+func withBackend(ip string, port uint32) backendPoolOption {
 	return func(o *isovalentv1alpha1.LBBackendPool) {
 		o.Spec.Backends = append(o.Spec.Backends, isovalentv1alpha1.Backend{
 			IP:     ip,
-			Port:   port,
+			Port:   int32(port),
 			Weight: ptr.To[uint32](1),
 		})
 	}
